Extract lookupListOfStrokesAndDicts.ts
