+++ src/components/StrokesForWords.js
@@ -1,13 +1,10 @@
 import * as React from 'react'
 import { Component } from 'react';
 import { SOURCE_NAMESPACES } from '../constant/index.js';
-import { AffixList } from '../utils/affixList';
-import rankOutlines from '../utils/transformingDictionaries/rankOutlines/rankOutlines';
 import splitBriefsIntoStrokes from './../utils/splitBriefsIntoStrokes';
+import lookupListOfStrokesAndDicts from "../utils/lookupListOfStrokesAndDicts";
 
 import misstrokesJSON from '../json/misstrokes.json'
-import getModifiedWordOrPhraseForLookup from '../utils/getModifiedWordOrPhraseForLookup';
-import createListOfStrokes from '../utils/createListOfStrokes';
 import getStenoDiagram from "../pages/lessons/utilities/getStenoDiagram";
 import getMapBriefsFn from "../pages/lessons/utilities/getMapBriefsFn";
 
@@ -211,71 +208,4 @@ class StrokesForWords extends Component {
   }
 }
 
-function lookupListOfStrokesAndDicts(phrase, globalLookupDictionary, affixList = AffixList.getSharedInstance()) {
-  let lookupText = phrase;
-  let modifiedWordOrPhrase = lookupText.slice();
-  modifiedWordOrPhrase = getModifiedWordOrPhraseForLookup(phrase);
-
-  let listOfStrokesAndDicts = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-
-  if (phrase === "{") {
-    modifiedWordOrPhrase = "{^}" + lookupText;
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary));
-  }
-  if (phrase === "}") {
-    modifiedWordOrPhrase = lookupText + "{^}";
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary));
-  }
-  // if (phrase === "[") { listOfStrokesAndDicts = listOfStrokesAndDicts.concat(createListOfStrokes("{^}" + lookupText, globalLookupDictionary)); }
-  // if (phrase === "]") { listOfStrokesAndDicts = listOfStrokesAndDicts.concat(createListOfStrokes(lookupText + "{^}", globalLookupDictionary)); }
-
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = "{^}" + lookupText + "{^}";
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = "{^}" + lookupText;
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = lookupText + "{^}";
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = "{^" + lookupText + "^}";
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = "{^" + lookupText + "}";
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = "{" + lookupText + "^}";
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = "{" + lookupText + "}";
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-  if (listOfStrokesAndDicts.length === 0) {
-    modifiedWordOrPhrase = lookupText.trim();
-    let listOfStrokesAndDictsWithSuppressedSpaces = createListOfStrokes(modifiedWordOrPhrase, globalLookupDictionary);
-    listOfStrokesAndDicts = listOfStrokesAndDicts.concat(listOfStrokesAndDictsWithSuppressedSpaces);
-  }
-
-  listOfStrokesAndDicts = rankOutlines(listOfStrokesAndDicts, misstrokesJSON, modifiedWordOrPhrase, affixList);
-
-  return [listOfStrokesAndDicts, modifiedWordOrPhrase];
-}
-
 export default StrokesForWords;
-export {
-  lookupListOfStrokesAndDicts
-};
